refactor: extract segment to a dedicated file.
